Add tests for delegate-effects
This commit adds tests for `delegate-effects.js` and uses the tracker
introduced in the previous commit.

In adding tests for `delegate-effects.js` I realized the previous
commit may not be necessary since `delegate-effects` already does what
it's doing... I can submit another commit to delete the new `yield-effect` rule
and keep the tests I added in this commit.